+++ internal/pkg/api_commons/FetchXML.go
