+++ pkg/lfm_api/UserGetRecentTracks.go
@@ -1,7 +1,7 @@
 package lfm_api
 
 import (
-	"github.com/dozro/tawny/pkg/api_commons"
+	"github.com/dozro/tawny/internal/pkg/api_commons"
 	"github.com/dozro/tawny/pkg/lfm_types"
 )
