+++ pkg/musicbrainz_api/lookup.go
@@ -3,7 +3,7 @@ package musicbrainz_api
 import (
 	"fmt"
 
-	"github.com/dozro/tawny/pkg/api_commons"
+	"github.com/dozro/tawny/internal/pkg/api_commons"
 	"github.com/dozro/tawny/pkg/musicbrainz_types"
 )
